+++ examples/web/script/main.js
@@ -4,11 +4,11 @@ const
 urlEl.value = "https://www.youtube.com/watch?v=4zRK0t4caOg"
 const
   parseVideo = () => {
-    const vidData = videoJson(extractVideo(urlEl.value, "https://api.allorigins.win/raw?url="))
+    const vidData = videoJson(extractVideo(urlEl.value, "https://corsproxy.io/?"))
     outEl.innerHTML = vidData
   },
   parseChannel = () => {
-    const vidData = channelJson(extractChannel(urlEl.value, "home", "https://api.allorigins.win/raw?url="))
+    const vidData = channelJson(extractChannel(urlEl.value, "home", "https://corsproxy.io/?"))
     outEl.innerHTML = vidData
   }
 document.getElementById("parseVideo").addEventListener("click", parseVideo)